Reorganize code generation and support generation from function definitions
